requirements.txt
django
djangorestframework
djoser
django-jazzmin
psycopg2-binary